refactor: Split project into modules
