+++ src/main.rs
@@ -1,18 +1,11 @@
 use std::sync::Arc;
 
-use serde::{Deserialize, Serialize};
-
-use axum::{
-    extract::{Path, State},
-    routing::get,
-    Router,
-};
-
-use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
-
-use uuid::Uuid;
+use axum::{routing::get, Router};
 
 use dotenvy::dotenv;
+use sqlx::postgres::PgPoolOptions;
+
+mod api;
 
 #[tokio::main]
 async fn main() {
@@ -46,8 +39,7 @@ async fn main() {
 
     let app = Router::new()
         .route("/", get(view_hw))
-        .route("/posts", get(list_posts))
-        .route("/posts/{post_id}", get(view_post))
+        .nest("/api", api::get_routes())
         .with_state(Arc::new(db_pool));
 
     let addr = std::env::var("HOST").unwrap_or("0.0.0.0:3000".to_string());
@@ -58,41 +50,4 @@ async fn main() {
 
 async fn view_hw() -> &'static str {
     "Hello from Axum, World!"
-}
-
-#[derive(sqlx::FromRow, Serialize, Deserialize)]
-struct Post {
-    uuid: Uuid,
-    slug: String,
-    title: String,
-    body: String,
-    is_published: bool,
-    #[serde(with = "time::serde::rfc3339")]
-    date_created: time::OffsetDateTime,
-    // TODO: Serialize this too with time::serde::rfc3339, tricky since Option :|
-    date_published: Option<time::OffsetDateTime>,
-}
-
-async fn list_posts(State(db_p): State<Arc<Pool<Postgres>>>) -> String {
-    // This query is a total mess due to sqlx being unable to determine the nullability of fields of a view
-    // See https://github.com/launchbadge/sqlx/issues/3192#issuecomment-2807790647
-    let out = sqlx::query_as!(Post, r#"SELECT * FROM post_view"#)
-        .fetch_all(&*db_p)
-        .await
-        .expect("couldn't query posts");
-
-    serde_json::to_string_pretty(&out).unwrap()
-}
-
-async fn view_post(Path(post_id): Path<Uuid>, State(db_p): State<Arc<Pool<Postgres>>>) -> String {
-    let out = sqlx::query_as!(
-        Post,
-        r#"SELECT * FROM post_view WHERE "uuid!" = $1::uuid"#,
-        post_id
-    )
-    .fetch_one(&*db_p)
-    .await
-    .expect("couldn't query posts");
-
-    serde_json::to_string_pretty(&out).unwrap()
 }
